% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateMGS.R
\name{CalculateMGS}
\alias{CalculateMGS}
\title{Calculate the mean group size per sample.}
\usage{
CalculateMGS(data, species, threshold, sampleName, preciselyCountedValue)
}
\arguments{
\item{data}{A FData object.}

\item{species}{A species name.}

\item{threshold}{A number of up to which the mean group size is calculated from a sample.
Otherwise, mean group size is calculated from all samples.}

\item{sampleName}{A vector of sample names in the same order as FDataYGDS object.}

\item{preciselyCountedValue}{A character value.}
}
\value{
return a vector of mean group size per sample.
}
\description{
Calculate the mean group size per sample.
}
