% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateSamplingEffort.R
\name{CalculateSamplingEffort}
\alias{CalculateSamplingEffort}
\title{Calculate sampling effort of sample for each replicate}
\usage{
CalculateSamplingEffort(data)
}
\arguments{
\item{data}{A FData object}
}
\value{
Return data.frame object with sampling effort of sample for each replicate
}
\description{
Calculate sampling effort of sample for each replicate
}
