% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateFData.R
\name{CreateFData}
\alias{CreateFData}
\title{Create FData object by formatting dataframe and adding replicate number according to samping effort.}
\usage{
CreateFData(species, number, preciselyCounted, distance, sample, subSample,
  order, samplingEffort)
}
\arguments{
\item{species}{A vector of species name.}

\item{number}{A vector of number of observed individuals.}

\item{preciselyCounted}{A vector indicating when the observed individuals were precisely counted.}

\item{distance}{A vector of perpendicular distance from the transect to the observation (individual or group).}

\item{sample}{A vector of sample name (site, year, etc).}

\item{subSample}{A vector of subsample name used to be aggregate to sample level.}

\item{order}{A vector using to define replicates number.}

\item{samplingEffort}{A vector of sampling effort for each replicate: number of m walked by the observer for the replicat}
}
\value{
Return a FData object: data.frame object with formated column names.
Replicate numbers are given according to sampling effort. Replicate numbers are ordered acoording to sampling effort.
First replicate numbers have the highest sampling effort whereas the lastest are the smallest.
Use CalculateSamplingEffort function to see sampling effort values per replicate.
}
\description{
Create FData object by formatting dataframe and adding replicate number according to samping effort.
}
\examples{
In progess
}
