% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateFDataYGDS.R
\name{CreateFDataYGDS}
\alias{CreateFDataYGDS}
\title{Create a FDataYGDS object to calculate IKA (Encounter Rate),and estimate abundance using gdistsamp function (package unmarked).}
\usage{
CreateFDataYGDS(data, siteCovs = NULL, samplingEffort, nbReplicat = 12,
  stripWidth = 80, DistanceClass = 5, unitsIn = "m", survey = "line",
  species = "Cebus", groupNumber = TRUE, naRmNumber = FALSE,
  naRmDistance = FALSE, missingValuesNumber = "mean",
  missingValuesDistance = "mean", ...)
}
\arguments{
\item{data}{A FData object.}

\item{siteCovs}{A data.frame object with covariates. Warning !!! Sample name in line must be ordered as FData object.}

\item{samplingEffort}{A data.frame object with sampling effort created by CalculateSamplingEffort function.}

\item{nbReplicat}{A replicate number. Default value is 12.}

\item{stripWidth}{A distance value used to define stripwidth and truncate distance. Default value is 80.}

\item{DistanceClass}{A distance value used as distance interval of classes. Default value is 5.}

\item{unitsIn}{Either "m" or "km" defining the measurement units for both stripWidth and DistanceClass.
See ? unmarkedFrameDS function (package unmarked). Default value is "m".}

\item{survey}{Either "point" or "line" for point- and line-transects.
See ? unmarkedFrameDS function (package unmarked).}

\item{species}{Chosen species. Default value is "line".}

\item{groupNumber}{Logical. If TRUE, abundance is calculated using groups as observations.
 If FALSE, abundance is calculated using individuals as observations.
All values of 'number' variable of FData object are replaced by 1, including NA. Default value is TRUE.}

\item{naRmNumber}{Logical. If TRUE, observations which are not species number values (NA) are removed. Default value is FALSE.}

\item{naRmDistance}{Logical. If TRUE, observations which are not distance values (NA) are removed. Default value is FALSE.}

\item{missingValuesNumber}{Either method "mean" or "random". See details. Default value is "mean".}

\item{missingValuesDistance}{Either method "mean" or "random". See details. Default value is "mean".}

\item{...}{other parameters of the unmarkedFrameGDS function(unmarked package)}
}
\value{
Return a list with a FData object, a array with observations,
and an unmarkedFrameGDS object which can used in gdistsamp function (package unmarked).
}
\description{
Create a FDataYGDS object to calculate IKA (Encounter Rate),and estimate abundance using gdistsamp function (package unmarked).
}
\details{
Method "mean" replaces a NA value by the mean value of the sample.
Method "random" replaces a NA value by a value drawn randomly from the sample.
}
