% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObsNumberPerSpecies.R
\name{ObsNumberPerSpecies}
\alias{ObsNumberPerSpecies}
\title{Calculate observation number per species.}
\usage{
ObsNumberPerSpecies(data)
}
\arguments{
\item{data}{A FData object.}
}
\value{
return a vector with observation number per species.
}
\description{
Calculate observation number per species.
}
\examples{
In progress
}
