% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotObservationPerDistanceClass.R
\name{PlotObservationPerDistanceClass}
\alias{PlotObservationPerDistanceClass}
\title{Represent observation number by distance class.}
\usage{
PlotObservationPerDistanceClass(data, ...)
}
\arguments{
\item{data}{A FDataYGDS object created by CreateFDataYGDS function.}

\item{...}{other parameters of the ggplot function (ggplot2 package)}
}
\value{
Return a plot of observation number (in y) by distance class (in x).
}
\description{
Represent observation number by distance class.
}
