% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSampleNumberVsSamplingEffort.R
\name{PlotSampleNumberVsSamplingEffort}
\alias{PlotSampleNumberVsSamplingEffort}
\title{Represent total sampling effort according to sample number.}
\usage{
PlotSampleNumberVsSamplingEffort(data, seqReplicat)
}
\arguments{
\item{data}{A FData object.}

\item{seqReplicat}{A numeric vector with replicate numbers}
}
\value{
Return a plot with the sampling effort in y and the sample number in x. Lables are the replicate number.
}
\description{
Represent total sampling effort according to sample number.
}
