% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReplaceSpeciesName.R
\name{ReplaceSpeciesName}
\alias{ReplaceSpeciesName}
\title{Replace species names by other term}
\usage{
ReplaceSpeciesName(speciesNames, taxonReplacement, speciesVector)
}
\arguments{
\item{speciesNames}{A vector of taxon names that are to be replaced}

\item{taxonReplacement}{A taxon name use to replace taxon names}

\item{speciesVector}{A vector of species names provided by FData object}
}
\value{
Return a vector of species names to replace species variable of dataframe created by CreateFData function
}
\description{
Replace species names by other term
}
\examples{
In progress
}
