% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReplaceValue.R
\name{ReplaceValue}
\alias{ReplaceValue}
\title{Replace NA values}
\usage{
ReplaceValue(data, colName, method)
}
\arguments{
\item{data}{A FData object}

\item{colName}{A column name}

\item{method}{Either "mean" or "random". Method "mean" replaces a NA value by the mean value of the sample.
Method "random" replaces a NA value by a value drawn randomly from the sample.}
}
\value{
Return a vector completed.
}
\description{
Replace NA values
}
\examples{
In progress
}
