% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReplicatePerSubSample.R
\name{ReplicatePerSubSample}
\alias{ReplicatePerSubSample}
\title{Give the replicate number per subsample}
\usage{
ReplicatePerSubSample(data)
}
\arguments{
\item{data}{A FData object}
}
\value{
return a vector with the replicate number per subsample
}
\description{
Give the replicate number per subsample
}
\examples{
In progess
}
