% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpeciesDensityEstimation.R
\name{SpeciesDensityEstimation}
\alias{SpeciesDensityEstimation}
\title{Estimate species density using gdistsamp function (package unmarked).}
\usage{
SpeciesDensityEstimation(data, modelGDS = "NB", methodGDS = "BFGS",
  keyfunGDS = "halfnorm", ...)
}
\arguments{
\item{data}{A FDataYGDS object.}

\item{modelGDS}{Either "P" or "NB" for the Poisson and negative binomial models of abundance. "NB" by default. See ? gdistsamp.}

\item{methodGDS}{Optimization method used by optim. "BFGS" by default.}

\item{keyfunGDS}{One of the following detection functions: "halfnorm", "hazard", "exp", or "uniform".
"halfnorm" by default. See ? gdistsamp.}

\item{...}{This function used CreateDataYGDS function with default arguments. See ? CreateDataYGDS.}
}
\value{
Return a list with an object of class unmarkedFitGD, a vector with IKA (Encounter rate for 10 km),
 a vector with abundance in km^{2} estimated by ranef function (package unmarked).
}
\description{
Estimate species density using gdistsamp function (package unmarked).
}
