% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TruncatedObservation.R
\name{TruncatedObservation}
\alias{TruncatedObservation}
\title{Calculate the number of truncated observation per sample.}
\usage{
TruncatedObservation(data)
}
\arguments{
\item{data}{A FDataYGDS object created by CreateFDataYGDS function}
}
\value{
Return the number of truncated observation per sample.
}
\description{
Calculate the number of truncated observation per sample.
}
\examples{
In progress
}
