% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotIkaDensity.R
\name{PlotIkaDensity}
\alias{PlotIkaDensity}
\title{Represent relatioship between IKA (i.e., Encounter rate, 10km walked length) and
estimated density.}
\usage{
PlotIkaDensity(data)
}
\arguments{
\item{data}{A FDataYGDS object created by CreateFDataYGDS function.}
}
\value{
Return a plot of estimated density (in x) and IKA (in x).
}
\description{
Represent relatioship between IKA (i.e., Encounter rate, 10km walked length) and
estimated density.
}
\examples{
In progress
}
